requirements.txt
hazm==0.7.0
wordcloud==1.8.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
python-bidi==0.4.2
arabic-reshaper==2.1.3